Fix DNS rebinding in webhook dispatch, MFA secret storage, and SQL cast
- Webhook::send(): pin cURL to the pre-validated IP via CURLOPT_RESOLVE
  to prevent DNS rebinding TOCTOU between the SSRF gethostbyname() check
  and the actual HTTP POST (same class of bug fixed in JWT/SSO last commit)
- dispatch_webhooks.php: cast BATCH_SIZE to int before SQL interpolation
  to make the type intent explicit (constant was safe but inconsistent)
- AuthController: defer TOTP secret DB write until after the user
  successfully verifies a code; secret is held in session only during
  setup, so abandoned MFA enrollments leave no orphan secret in the DB

+++ aegis/scripts/dispatch_webhooks.php
@@ -47,7 +47,7 @@ $deliveries = Database::fetchAll(
       WHERE status = 'pending'
         AND next_retry_at <= NOW()
       ORDER BY next_retry_at ASC
-      LIMIT " . BATCH_SIZE
+      LIMIT " . (int) BATCH_SIZE
 );
 
 $totalDelivered = 0;
